+++ src/redux/actions/surveyActions.js
@@ -3,7 +3,8 @@ import {
   CREATE_SURVEY,
   UPDATE_SURVEY,
   GET_SURVEY_BY_ID,
-  DELETE_SURVEY 
+  DELETE_SURVEY,
+  SAVE_SURVEY_ANSWER 
 } from '../consts';
 
 export const createSurvey = payload => ({
@@ -23,5 +24,10 @@ export const getSurveyById = payload => ({
 
 export const deleteSurvey = payload => ({
   type: DELETE_SURVEY,
+  payload
+});
+
+export const saveSurveyAnswer = payload => ({
+  type: SAVE_SURVEY_ANSWER,
   payload
 });